pyproject.toml
[project]
name = "events-face"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "rest-framework-simplejwt>=0.0.2",
    "ruff>=0.14.9",
]
[tool.ruff.lint]
extend-select = ["I"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
